add create, add, read and delete function
